refactor(app_runtime): Board handler を board.rs に抽出 (SPEC-2077 Phase A)
- crates/gwt/src/app_runtime.rs を app_runtime/mod.rs にディレクトリ化
- crates/gwt/src/app_runtime/board.rs を新規作成し、BoardPostRequest struct と post_board_entry_events impl extension、sanitize_board_list helper を移行
- mod.rs に mod board; 宣言と pub(crate) use board::BoardPostRequest を追加
- main.rs の include_str!("app_runtime.rs") を include_str!("app_runtime/mod.rs") へ更新
- god module 解体 phase chain の最初の phase (Phase A)、 1 module 抽出 + dispatch 経路 1 箇所のみ変更で blast radius 最小

Verification:
- cargo test -p gwt-core -p gwt: 全 pass (SPEC-1974 Board test 群が新 module 経由で全件 pass)
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs 7445→7293 行 (152 行減)、 board.rs 新 177 行
